requirements.txt
# Requirements for Renewable Energy Dispute Characterization Project
# Dependencies for running the data pipeline

# Data Processing & Analysis
pandas>=1.5.0
numpy>=1.20.0
scipy>=1.7.0
pickle5; python_version < "3.8"

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
folium>=0.12.0
wordcloud>=1.8.0

# Geospatial Analysis
geopandas>=0.12.0
shapely>=1.8.0

# LLM Integration
anthropic>=0.3.0
openai>=1.0.0
instructor>=0.4.0
pydantic>=2.0.0

# Web Scraping & Content Processing
requests>=2.28.0
unstructured[all-docs]>=0.10.0
modal>=0.50.0

# Web Applications
streamlit>=1.20.0
streamlit-js-eval>=0.1.0

# Utilities & Progress
tqdm>=4.64.0
backoff>=2.2.0
python-dotenv>=0.19.0

# AWS S3 Data Management
boto3>=1.26.0

# Natural Language Processing
nltk>=3.8.0

# Google Services & Data
gspread>=5.0.0
google-generativeai>=0.3.0

# Development & Testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
